requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
httplib2==0.22.0
idna==3.10
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ShopifyAPI==12.7.0
six==1.17.0
tabulate==0.9.0
tzdata==2024.2
urllib3==2.3.0
python-dotenv
